Tehty pieniä korjauksia tiedostorakenteeseen
