+++ src/main/java/fi/academy/todomon/WebSecurityConfig.java
@@ -1,5 +1,6 @@
-package fi.academy.todomon.config;
+package fi.academy.todomon;
 
+import fi.academy.todomon.CustomizedAuthenticationSuccessHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -7,15 +8,7 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.NoOpPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.provisioning.JdbcUserDetailsManager;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import javax.sql.DataSource;
 
@@ -70,18 +63,18 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
         }
     }
 
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth) {
-        try {
-            auth
-                    .inMemoryAuthentication()
-                    .withUser("user").password("user").roles("USER")
-                    .and()
-                    .withUser("admin").password("admin").roles("ADMIN");
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
+//    @Autowired
+//    public void configureGlobal(AuthenticationManagerBuilder auth) {
+//        try {
+//            auth
+//                    .inMemoryAuthentication()
+//                    .withUser("user").password("user").roles("USER")
+//                    .and()
+//                    .withUser("admin").password("admin").roles("ADMIN");
+//        } catch (Exception e) {
+//            e.printStackTrace();
+//        }
+//    }
 }
