+++ src/main/resources/templates/annintesti.html
@@ -4,18 +4,16 @@
 <head>
     <meta charset="UTF-8">
     <title>hello security</title>
+    <link rel="stylesheet"
+          href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
 </head>
 <body>
-<h1 th:inline="text">Hello [[${#httpServletRequest.remoteUser}]]!</h1>
+<h1 th:text="'Hello ' + ${T(org.springframework.security.core.context.SecurityContextHolder)
+      .context.authentication.principal.username} + '!'"></h1>
 <form action="/logout" method="post">
     <input type="hidden"
            th:name="${_csrf.parameterName}" th:value="${_csrf.token}"/>
     <input type="submit" value="Log Out">
-
-    // alempi koodi hakee myös käyttäjänimen
-    <p>Hello, <span th:text="${
-    T(org.springframework.security.core.context.SecurityContextHolder)
-      .context.authentication.principal.username}"></span>!</p>
 </form>
 </body>
 </html>
