+++ src/main/resources/templates/home.html
@@ -3,10 +3,12 @@
       xmlns:th="http://www.thymeleaf.org"
       xmlns:sec="http://www.thymeleaf.org/thymeleaf-extras-springsecurity3">
 <head>
-    <title>Security with Spring Boot</title>
+    <title>ToDoMon - gotta do 'em all</title>
+    <link rel="stylesheet"
+          href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"/>
 </head>
 <body>
-<h1>Hello, This is Home page!</h1>
+<h1>Welcome to ToDoMon</h1>
 <a style="color: blue" th:href="@{/user}">User Page</a>
 <br />
 <a style="color: blue" th:href="@{/admin}">Admin Page</a>
